Make small changes in Users page

+++ PizzaShop.Web/Controllers/UserController.cs
@@ -214,8 +214,15 @@ public class UserController : Controller
     }
 
     [HttpPost]
-    public IActionResult DeleteUser(int id)
+    [Route("User/DeleteUser/{id}/{roleId}")]
+    public IActionResult DeleteUser(int id, int roleId)
     {
+        if (roleId == 1)
+        {
+            TempData["ToastrMessage"] = "You don't have access to delete the Admin user";
+            TempData["ToastrType"] = "error";
+            return RedirectToAction(nameof(UserList));
+        }
         _userService.DeleteUser(id);
         TempData["ToastrMessage"] = "User Deleted Successfully";
         TempData["ToastrType"] = "success";
